+++ Unity_EyapLibrary/Packages/com.maellacour.eyap-library/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+Add the SceneManagement tools, so that you can load multiple scenes efficiently, based on a scriptable object requirements.
+
+### Added
+- SceneManagement: Scene loader callable from anywhere (static class).
+- SceneManagement: SceneSwitcher as a persistent Singleton.
+- SceneManagement: Initialization Loader, that loads scene defined in a SO.
+
+
 ## [1.1.0] - 2023-09-13
 ### Added
 - Editor: Method to export all localization files as csv. Infortunately, that add the Localization package as dependency.